backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aioredis[hiredis]==2.0.1
    # via -r requirements.in
alembic==1.14.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via
    #   aioredis
    #   asyncpg
asyncpg==0.30.0
    # via -r requirements.in
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.4
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
google-api-core==2.22.0
    # via google-api-python-client
google-api-python-client==2.151.0
    # via -r requirements.in
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via google-api-core
greenlet==3.1.1
    # via sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==3.0.0
    # via aioredis
httpcore==1.0.6
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
mako==1.3.6
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via gunicorn
proto-plus==1.25.0
    # via google-api-core
protobuf==5.28.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via -r requirements.in
pyparsing==3.2.0
    # via httplib2
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.17
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
starlette==0.41.2
    # via fastapi
typer==0.13.0
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   aioredis
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich
    #   sqlalchemy
    #   typer
    #   uvicorn
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.32.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==14.0
    # via uvicorn
